[IMP] tracking_manager: Test compatibility
Previously the bank_ids field was used but account adds messages in partner when
a related bank account is created/modified/deleted, causing tests to fail.

The user_ids field of res.partner is now used to avoid the side effect in tests
and the fake model is removed.

TT50571